\name{svSweave-package}
\alias{svSweave-package}
\alias{svSweave}
\docType{package}

\title{ Sweave, Knitr and R Markdown Companion Functions }
\description{
  The SciViews svSweave package provides additional function to use Sweave with
  LyX and SciViews-LyX. Also function to reference tables, figures or equations inside a R markdown file are added with v1.0.
}
\details{
  \tabular{ll}{
    Package: \tab svSweave\cr
    Type: \tab Package\cr
    Version: \tab 1.0.0\cr
    Date: \tab 2018-08-17\cr
    License: \tab GPL 2 or above, at your convenience\cr
  }
  % TODO: add description of main functions here. Also add examples
}

\author{
  Philippe Grosjean

  Maintainer: Ph. Grosjean <phgrosjean@sciviews.org>
}

\keyword{ package }
\keyword{ utilities }

\concept{ GUI API sweave LyX litterate programming }
