% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_labelling.R
\name{new_labelling}
\alias{new_labelling}
\alias{newLabelling}
\alias{fig}
\alias{tab}
\alias{eq}
\title{Reference figures, tables and equations in R Markdown documents}
\usage{
new_labelling(
  type = c("arabic", "roman"),
  prefix = "",
  string = "_**Figure ####:** @@@@_"
)

newLabelling(
  type = c("arabic", "roman"),
  prefix = "",
  string = "_**Figure ####:** @@@@_"
)

fig(label, text, reset = FALSE)

tab(label, text, reset = FALSE)

eq(label, text, reset = FALSE)
}
\arguments{
\item{type}{The type of numbering (arabic or roman).}

\item{prefix}{The text to use to prefix the number (nothing, by default).}

\item{string}{The string prototyping the legend, with \verb{@@} being a
placeholder for the text, \verb{####} as a placeholder for the number, or
\verb{\%\%\%\%} as a placeholder for numbered examples (used by \code{\link[=eq]{eq()}}).}

\item{label}{A short string uniquely identifying the item within the
collection.}

\item{text}{The text for the legend (can contain R Markdown formatting). In
the case of \code{\link[=eq]{eq()}}, this is the LaTeX code for the equation itself.}

\item{reset}{If \code{TRUE}, the collection is reset. Useful if you want to
restart numbering at the beginning of each chapter.}
}
\value{
The item labelling (legend, ...) if `text =`` is provided, or the number
associated with the item otherwise. \code{\link[=new_labelling]{new_labelling()}} creates a new labelling
function, which has the same arguments as \code{\link[=fig]{fig()}}.
}
\description{
These functions return closures that allow for constructing a series of
numbered items and to reference them. The number is create the first time a
label is encountered, and provided again for further use of the same label.
}
\details{
A new labelling type is created using \code{\link[=new_labelling]{new_labelling()}} which is a function
factory (a function that creates functions).
}
\examples{
# Note that those instructions are supposed to be included in R Markdown
# documents using the inline R expression syntax: `r fig("label", "<text>")`.

# Say you have a first figure labelled 'first-plot'.
# The first call creates its legend, and the second one references it
fig("first-plot", "An example of a gpplot2 scatterplot.")
fig("first-plot")

# A second figure labelled 'a-picture', but with reference appearing first
fig("a-picture")
# The second call creates the legend for the picture
fig("a-picture", "This is my most beautiful picture")

# Same for a table labelled 'anova-table'
tab("anova-table", "ANOVA for my wonderful dataset.")
# ... and further in the text, a reference to that table:
tab("anova-table")
}
\author{
Philippe Grosjean
}
\concept{automatic numbering of items in documents}
\keyword{utilities}
