% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svSweave-package.R
\docType{package}
\name{svSweave-package}
\alias{svSweave}
\alias{svSweave-package}
\title{svSweave: 'SciViews' - 'Sweave', 'Knitr' and R Markdown Companion Functions}
\description{
Functions to enumerate and reference figures, tables and equations
  in R Markdown documents that do not support these features (thus not
  'bookdown' or 'quarto'. Supporting functions for using 'Sweave' and 'Knitr'
  with 'LyX'.
}
\details{
The SciViews 'svSweave' package provides additional function to use Sweave or
knitr with LyX and SciViews-LyX. There are also functions to reference
tables, figures or equations inside a R markdown.
}
\section{Important functions}{

\itemize{
\item \code{\link[=fig]{fig()}}, \code{\link[=tab]{tab()}} or \code{\link[=eq]{eq()}} to enumerate and reference figures, tables and
equations in R Markdown documents.
\item \code{\link[=new_labelling]{new_labelling()}} to create new items iterators.
\item \code{\link[=clean_lyx]{clean_lyx()}}, \code{\link[=weave_lyx]{weave_lyx()}}, \code{\link[=knit_lyx]{knit_lyx()}}, \code{\link[=tangle_lyx]{tangle_lyx()}}, \code{\link[=purl_lyx]{purl_lyx()}}
to clean up, weave or tangle Sweave files produced by LyX with the SciViews
Sweave or Knitr modules.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SciViews/svSweave}
  \item \url{https://www.sciviews.org/svSweave/}
  \item Report bugs at \url{https://github.com/SciViews/svSweave/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

}
\keyword{internal}
