\name{svSweave-package}
\alias{svSweave-package}
\alias{svSweave}
\docType{package}

\title{ SciViews GUI API - Sweave functions }
\description{
  The SciViews svSweave package provides additional function to use Sweave with
  LyX and SciViews-LyX.
}
\details{
  \tabular{ll}{
    Package: \tab svSweave\cr
    Type: \tab Package\cr
    Version: \tab 0.9-6\cr
    Date: \tab 2013-01-02\cr
    License: \tab GPL 2 or above, at your convenience\cr
  }
  % TODO: add description of main functions here. Also add examples
}

\author{
  Philippe Grosjean

  Maintainer: Ph. Grosjean <phgrosjean@sciviews.org>
}

\keyword{ package }
\keyword{ utilities }

\concept{ GUI API sweave LyX litterate programming }
