% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svSweave-package.R
\docType{package}
\name{svSweave-package}
\alias{svSweave-package}
\title{SvSweave - Sweave, Knitr and R Markdown Companion Functions}
\description{
The SciViews 'svSweave' package provides additional function to use Sweave or
knitr with LyX and SciViews-LyX. There are also functions to reference
tables, figures or equations inside a R markdown.
}
\section{Important functions}{

\itemize{
\item \code{\link[=fig]{fig()}}, \code{\link[=tab]{tab()}} or \code{\link[=eq]{eq()}} to enumerate and reference figures, tables and
equations in R Markdown documents.
\item \code{\link[=new_labelling]{new_labelling()}} to create new items iterators.
\item \code{\link[=clean_lyx]{clean_lyx()}}, \code{\link[=weave_lyx]{weave_lyx()}}, \code{\link[=knit_lyx]{knit_lyx()}}, \code{\link[=tangle_lyx]{tangle_lyx()}}, \code{\link[=purl_lyx]{purl_lyx()}}
to clean up, weave or tangle Sweave files produced by LyX with the SciViews
Sweave or Knitr modules.
}
}

